Adapt microgrid info tests to the common client layout
Split the imported test into a pure-type test (`tests/microgrid/`) and a
protobuf conversion test (`tests/microgrid/proto/v1alpha8/`), mirroring
the source tree, and import from `frequenz.client.common`.

The conversion test patches the converters in their new location and
mocks the local `datetime_from_proto` helper instead of
`conversion.to_datetime`.
