+++ tests/microgrid/proto/v1alpha8/__init__.py
@@ -1,0 +1,4 @@
+# License: MIT
+# Copyright © 2026 Frequenz Energy-as-a-Service GmbH
+
+"""Tests for the microgrid.proto.v1alpha8 package."""
